config/requirements.txt

# Authentication App Requirements
# File: apps/auth/requirements.txt

# Core Django
Django==5.0.0

python-decouple==3.8

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg[binary]==3.2.13



# Security
PyJWT==2.8.0

# Email
django-ses==3.5.2  # Optional: for AWS SES

# Utilities
python-dateutil==2.8.2

# CORS Headers (for React frontend)
django-cors-headers==4.3.1

# Environment variables
python-dotenv==1.0.0